Fix Lliam deduplication issues. (#193)
When dedupe_reservoir was run, it was not accounting for title_abbrev
fields that were already appended with a number. This change compares
the field values without any appended numbers, and then appends a new
number based on the count. Initially I wanted to just pick up where
the count left off. However, it turned out to be possible to generate
duplicate titles on subsequent runs, so there was no way to determine
what the 'original' duplicate was.